requirements.txt
absl-py==2.0.0
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
apturl==0.5.2
astor==0.8.1
asttokens==2.3.0
async-timeout==4.0.2
atom3d==0.2.6
attrdict==2.0.1
attrs==23.1.0
backcall==0.2.0
bce-python-sdk==0.8.83
bcrypt==3.1.7
beautifulsoup4==4.8.2
biopython==1.81
blinker==1.6.2
blosc2==2.0.0
botocore==1.29.135
Brlapi==0.7.0
Brotli==1.0.9
cachetools==5.3.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==3.1.0
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
Cython==3.0.2
datasets==2.11.0
dbus-python==1.2.16
debugpy==1.6.7.post1
decorator==4.4.2
defer==1.0.6
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
duplicity==0.8.12.0
easy-parallel==0.1.6
entrypoints==0.3
executing==1.2.0
fastapi==0.95.1
fasteners==0.14.1
ffmpy==0.3.0
filelock==3.12.0
Flask==2.3.1
Flask-Babel==2.0.0
freesasa==2.2.0.post3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gevent==22.10.2
geventhttpclient==2.0.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
gradio==3.27.0
gradio_client==0.1.3
greenlet==2.0.2
grpcio==1.59.0
gssapi==1.6.1
h11==0.14.0
h5py==3.9.0
html5lib==1.0.1
httpcore==0.17.0
httplib2==0.14.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==2.8
importlib-metadata==6.6.0
ipaclient==4.8.6
ipalib==4.8.6
ipaplatform==4.8.6
ipapython==4.8.6
ipykernel==6.25.1
ipython==8.12.2
isc==2.0
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==2.10.1
joblib==1.2.0
jsonline==0.2.1
jsonlines==3.1.0
jupyter_client==8.3.1
jupyter_core==5.3.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightning-utilities==0.9.0
lmdb==1.4.1
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.5
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.5
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multipledispatch==1.0.0
multiprocess==0.70.14
nest-asyncio==1.5.7
netaddr==0.7.19
netifaces==0.10.4
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
onnx==1.13.1
orjson==3.8.10
packaging==23.2
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlenlp==2.5.2
paddlepaddle-gpu==0.0.0.post112
pandas==2.0.1
paramiko==2.6.0
parso==0.8.3
pathos==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==3.10.0
ply==3.11
pox==0.3.3
ppft==1.7.6.7
progressbar==2.5
prompt-toolkit==3.0.39
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.19
pycryptodome==3.17
pycups==1.9.73
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrr==0.10.3
python-apt==2.0.1+ubuntu0.20.4.1
python-augeas==0.5.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
python-ldap==3.2.0
python-multipart==0.0.6
python-rapidjson==1.10
python-yubico==1.3.3
pytorch-lightning==2.0.8
pytz==2023.3
pyusb==1.0.2
pyxdg==0.26
PyYAML==6.0.1
qrcode==6.1
rarfile==4.0
regex==2023.5.5
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
responses==0.18.0
rich==13.3.4
rouge-chinese==1.0.3
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
SecretStorage==2.3.1
semantic-version==2.10.0
sentencepiece==0.1.98
seqeval==1.2.2
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
ssh-import-id==5.10
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
systemd-python==234
tables==3.8.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.1.0
torchaudio==0.13.1+cu117
torchmetrics==1.1.1
torchvision==0.14.1+cu117
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.1.0
tritonclient==2.33.0
ttystatus==0.38
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.15
usb-creator==0.3.7
uvicorn==0.21.1
visualdl==2.5.1
wadllib==1.3.3
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.2
Werkzeug==2.3.0
x2paddle==1.4.1
xkit==0.0.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
zope.event==4.6
zope.interface==6.0
